pyproject.toml
[tool.poetry]
name = "y-lab-testovoe"
version = "0.1.0"
description = ""
authors = ["Idel Fazletdinov <fvi-it@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.109.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
uvicorn = {extras = ["standart"], version = "^0.26.0"}
asyncpg = "^0.29.0"
typing-extensions = "^4.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
mypy = "^1.8.0"
flake8 = "^7.0.0"
pre-commit = "^3.6.0"
pytest-asyncio = "<0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
